+++ workflows/src/copy-template-machine/copy-template-machine.test.ts
